Move funcoes para o formStore
